pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "smp"
version = "0.1.0"
description = "Structural Memory Protocol — graph-based codebase intelligence for AI agents"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "SMP Team" }]

dependencies = [
    "msgspec>=0.19",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "neo4j>=5.0",
    "chromadb>=0.6",
    "httpx>=0.27",
    "tree-sitter>=0.24",
    "tree-sitter-python>=0.23",
    "tree-sitter-typescript>=0.23",
    "google-genai>=1.0",
    "python-dotenv>=1.0",
    "structlog>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.8",
    "mypy>=1.13",
]

[tool.hatch.build.targets.wheel]
packages = ["smp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "ANN"]
ignore = ["ANN101", "ANN102", "ANN401"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[project.scripts]
smp = "smp.cli:main"
